ARM: tegra: Clean up external memory controller nodes
The external memory controller should be sorted after the memory
controller to keep the ordering by unit-address intact. While at it,
rename the emc-timings and timing nodes to avoid including the RAM code
and clock frequency in their names. There is no requirement to do this,
so we can use simple enumerations instead.

Signed-off-by: Thierry Reding <treding@nvidia.com>