ARM: tegra: Clean up display controller nodes
The display controllers found on Tegra30 and Tegra114 are not exactly
compatible with those on Tegra20. While the programming model is very
similar, there are various properties that do vary between the chips.

Buffer alignment restrictions and capabilities are not the same for all
the chips, so make sure that Tegra30 and Tegra114 are not marked as
compatible with Tegra20.

While at it, also remove the parent clock because it always depends on
the output that's consuming data from the display controller.

Signed-off-by: Thierry Reding <treding@nvidia.com>