gki_defconfig: Enable tegra186
Enable ARCH_TEGRA and ARCH_TEGRA_186_SOC as builtin

MAILBOX is now automatically selected by the Tegra BPMP driver
so it's drops from the defconfig with these additions.

Tegra186 uses SRAM for it's initial Boot Power Management Processor
so enable that as well. The Tegra memory controller also depends on
CONFIG_MEMORY, neither of which can be built as modules. With these
changes, GENERIC_PHY drops as it is automatically selected.

Change-Id: I918da0933c9029759de0448bfe622b41d50ff8f8